Address PR review comments for plugin EP allocator adapter
- Fix OrtAllocator leak: wrap in Ort::Allocator for RAII ownership

- Add IsStreamAware/AllocOnStream with version-gated function pointer forwarding

- Move IAllocatorWrappingOrtAllocator to allocator.h alongside the existing Allocator adapter

- Add TODO comments for Ort::Allocator C++ API gap and GetStats follow-up
