+++ include/onnxruntime/ep/adapter/allocator.h
@@ -9,11 +9,62 @@
 
 #include <mutex>
 
+#include <gsl/gsl>
+
 #include "core/framework/allocator.h"
+#include "core/framework/plugin_ep_stream.h"
 
 namespace onnxruntime {
 namespace ep {
 namespace adapter {
+
+// Wraps an OrtAllocator* exposed by the C API as an IAllocator.
+// Takes ownership of the OrtAllocator* and releases it on destruction.
+class IAllocatorWrappingOrtAllocator final : public IAllocator {
+ public:
+  explicit IAllocatorWrappingOrtAllocator(gsl::not_null<OrtAllocator*> ort_allocator)
+      : IAllocator(*ort_allocator->Info(ort_allocator)), ort_allocator_(ort_allocator.get()) {}
+
+  void* Alloc(size_t size) override {
+    return ort_allocator_.Alloc(size);
+  }
+
+  void Free(void* p) override {
+    ort_allocator_.Free(p);
+  }
+
+  void* Reserve(size_t size) override {
+    return ort_allocator_.Reserve(size);
+  }
+
+  bool IsStreamAware() const override {
+    return SupportsAllocOnStream();
+  }
+
+  void* AllocOnStream(size_t size, Stream* stream) override {
+    // TODO: Replace direct function pointer access once Ort::Allocator exposes AllocOnStream/IsStreamAware.
+    if (SupportsAllocOnStream()) {
+      OrtAllocator* raw = ort_allocator_;
+      return raw->AllocOnStream(raw, size, static_cast<OrtSyncStream*>(stream));
+    }
+    return Alloc(size);
+  }
+
+ private:
+  static constexpr uint32_t kOrtAllocatorAllocOnStreamMinVersion = 23;
+
+  bool SupportsAllocOnStream() const {
+    const OrtAllocator* raw = ort_allocator_;
+    return raw->version >= kOrtAllocatorAllocOnStreamMinVersion && raw->AllocOnStream != nullptr;
+  }
+
+  // TODO: Consider adding GetStats() override. Requires parsing OrtKeyValuePairs from the C API
+  // into AllocatorStats; see GetStatsFromOrtAllocator() in allocator_adapters.cc for reference.
+
+  Ort::Allocator ort_allocator_;
+
+  ORT_DISALLOW_COPY_ASSIGNMENT_AND_MOVE(IAllocatorWrappingOrtAllocator);
+};
 
 /// <summary>
 /// A bridge class between the EP API OrtAllocator and an IAllocator implementation.
